% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding_functions.R
\name{a_sub_b}
\alias{a_sub_b}
\title{A is subset of B}
\usage{
a_sub_b(a, b)
}
\arguments{
\item{a}{first interval}

\item{b}{second interval}
}
\value{
TRUE if there is intersection, FALSE otherwise
}
\description{
A is subset of B
}
