% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding_functions.R
\name{a_subeq_b}
\alias{a_subeq_b}
\title{A is subset or equal to B}
\usage{
a_subeq_b(a, b)
}
\arguments{
\item{a}{first interval}

\item{b}{second interval}
}
\value{
TRUE if there is intersection, FALSE otherwise
}
\description{
A is subset or equal to B
}
