% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding_functions.R
\name{emb_fd}
\alias{emb_fd}
\title{Fodor embedding function}
\usage{
emb_fd(a, b)
}
\arguments{
\item{a}{interval}

\item{b}{interval}
}
\value{
Embedding value
}
\description{
Fodor embedding function
}
