% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding_functions.R
\name{emb_gd}
\alias{emb_gd}
\title{Godel embedding function}
\usage{
emb_gd(a, b)

emb_gd(a, b)

emb_gd(a, b)
}
\arguments{
\item{a}{interval}

\item{b}{interval}
}
\value{
Embedding value

Embedding value

Embedding value
}
\description{
Godel embedding function

Goguen embedding function

Rescher embedding function
}
