% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding_functions.R
\name{emb_gg}
\alias{emb_gg}
\title{Goguen embedding function}
\usage{
emb_gg(a, b)
}
\arguments{
\item{a}{interval}

\item{b}{interval}
}
\value{
Embedding value
}
\description{
Goguen embedding function
}
