% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding_functions.R
\name{emb_lk}
\alias{emb_lk}
\title{Lukasiewicz embedding function}
\usage{
emb_lk(a, b)
}
\arguments{
\item{a}{interval}

\item{b}{interval}
}
\value{
Embedding value
}
\description{
Lukasiewicz embedding function
}
