% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding_functions.R
\name{emb_rs}
\alias{emb_rs}
\title{Rescher embedding function}
\usage{
emb_rs(a, b)
}
\arguments{
\item{a}{interval}

\item{b}{interval}
}
\value{
Embedding value
}
\description{
Rescher embedding function
}
