% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding_functions.R
\name{intersection}
\alias{intersection}
\title{Intersection between two intervals}
\usage{
intersection(a, b, interval = FALSE)
}
\arguments{
\item{a}{first interval}

\item{b}{second interval}
}
\value{
TRUE if there is intersection, FALSE otherwise
}
\description{
Intersection between two intervals
}
