% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_functions.R
\name{sim_emb}
\alias{sim_emb}
\alias{sim_w}
\alias{sim_lk}
\alias{sim_fd}
\alias{sim_gd}
\alias{sim_gg}
\alias{sim_rs}
\title{Similarity given by embeddings}
\usage{
sim_emb(emb, a, b)

sim_w(a, b)

sim_w(a, b)

sim_lk(a, b)

sim_fd(a, b)

sim_gd(a, b)

sim_gg(a, b)

sim_rs(a, b)
}
\arguments{
\item{b}{}
}
\value{



}
\description{
Similarity given by embeddings

Similarity width
}
