% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_matrix.R
\name{sim_emb_bt_two_obj}
\alias{sim_emb_bt_two_obj}
\title{Similarity between two objects}
\usage{
sim_emb_bt_two_obj(obj1, obj2, emb_sim, agg = NULL)
}
\arguments{
\item{obj1}{Object defined by a vector with a pair number of elements.
Consecutive elements represent lower (odd elements) and upper (even elements)
bounds of an interval.}

\item{obj2}{The object to measure the similarity with obj1.}

\item{emb_sim}{Function that given two intervals a,b returns a value of similarity}
}
\value{
A vector with the similariy if agg is NULL and the value aggregated using agg otherwise
}
\description{
Compute the similarity between two vectors using similarity measures
based on embedding functions.
}
