% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_matrix.R
\name{sim_emb_matrix}
\alias{sim_emb_matrix}
\title{Matrix of similarities}
\usage{
sim_emb_matrix(data, emb_sim, agg, verbose = F)
}
\arguments{
\item{data}{Dataset where each row is one of the objects. The data must have
an even pair of variables where the i-th column such that i is an odd number
is the lower bound of the interval representing the variable and the i+1 column
the corresponding upper bound.}

\item{emb_sim}{Embedding similarity}

\item{agg}{Aggregation function e.g. sum, mean}

\item{verbose}{}
}
\value{

}
\description{
Given a dataset this computes a matrix of similarities using an embe
}
