% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding_functions.R
\name{similarity}
\alias{similarity}
\title{Rescher embedding function}
\usage{
similarity(a, b, method)
}
\arguments{
\item{a}{interval}

\item{b}{interval}
}
\value{
Embedding value
}
\description{
Rescher embedding function
}
