% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding_functions.R
\name{union}
\alias{union}
\title{Union between two intervals}
\usage{
union(a, b, interval = FALSE)
}
\arguments{
\item{a}{first interval}

\item{b}{second interval}
}
\value{

}
\description{
Union between two intervals
}
